Iteration #3 Completion no bios
Altered the catalog page to pull from the database but cannot actually
click on the item for more detail. Removed the foreign keys from User
Cart and Billing allowed a user to be able to alter their billing
account as well as create a new one. Bios page can be added and
implemented quickly